PM-40520 - Rename wire_v1 to data_v1 and clarify layer docs
`wire_v1` was ambiguous — there are three "wires" in this crossing
(plaintext payload, sealed opaque blob, outbound JSON) and the filename
gave no clue which one was versioned. Rename to `data_v1` to match the
type identity (`RegistrationOpenOrgInviteData` / `...V1`) and rewrite
the module doc to name the layer explicitly and contrast it with the
outer two wires.
