+++ crates/bitwarden-auth/src/registration/open_org_invite/mod.rs
@@ -1,22 +1,23 @@
 //! Open-organization-invite registration crossing.
 //!
 //! The app seals an invite context on registration-start submit and unseals it on the accept
-//! open-org-invite component after a successful registration-finish. This module owns the shared
-//! wire schema (`wire_v1`), the seal path (`seal`), and the unseal path (`unseal`).
+//! open-org-invite component after a successful registration-finish. This module owns the
+//! versioned plaintext payload (`data_v1`), the seal path (`seal`), and the unseal path
+//! (`unseal`).
 
+mod data_v1;
 mod seal;
 mod unseal;
-mod wire_v1;
 
 use bitwarden_crypto::{
     generate_versioned_sealable,
     safe::{DataEnvelopeNamespace, SealableData, SealableVersionedData},
 };
+use data_v1::RegistrationOpenOrgInviteDataV1;
 pub use seal::{
     OpenOrgInviteSealRequest, SealedEnvelopePair, SealedEnvelopePairError, SealedOpenOrgInvite,
 };
 use serde::{Deserialize, Serialize};
-use wire_v1::RegistrationOpenOrgInviteDataV1;
 
 generate_versioned_sealable!(
     RegistrationOpenOrgInviteData,
